Inclusão de docker-compose

+++ Stone.PSP.Domain/Entities/Payable.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; init; }
         public Guid TransactionId { get; init; }
         public decimal Value { get; init; }
-        public int Status { get; init; }
+        public byte Status { get; init; }
         public DateTime PaymentDate { get; init; }
     }
 }
